requirement.txt
Django==1.11
MySQL-python==1.2.5
Pillow==4.0.0
django-simple-captcha==0.5.5
